# api/openapi.yaml
basePath: /
definitions:
...
  domain.BudgetPeriod:
    enum:
    - daily
    - weekly
    - monthly
    - yearly
    - custom
    type: string
    x-enum-varnames:
    - BudgetDaily
    - BudgetWeekly
    - BudgetMonthly
    - BudgetYearly
    - BudgetCustom
  domain.BudgetRecord:
    properties:
      amountLimitMinor:
        description: AmountLimitMinor is the limit in the currency's minor units.
        type: integer
      categoryID:
        type: string
      createdAt:
        description: CreatedAt and UpdatedAt record projection timestamps.
        type: string
      currency:
        type: string
      deletedAt:
        description: DeletedAt marks a soft-deleted budget.
        type: string
      dueAt:
        description: DueAt optionally caps a custom budget.
        type: string
      id:
        description: ID identifies the budget.
        type: string
      name:
        type: string
      ownerID:
        type: string
      period:
        allOf:
        - $ref: '#/definitions/domain.BudgetPeriod'
        description: Period defines the budget window.
      sharedGroupID:
        type: string
      thresholdPercent:
        description: ThresholdPercent triggers an alert when this percentage is reached.
        type: integer
      updatedAt:
        type: string
    type: object
...
  domain.ExtractedExpense:
    properties:
      address:
        description: Address is the merchant's address as printed on the receipt,
          if any.
        type: string
      amount_minor:
        description: AmountMinor is the total amount in minor currency units.
        type: integer
      category_name:
        description: CategoryName is a free-text category guess.
        type: string
      currency:
        description: Currency is the ISO 4217 currency code.
        type: string
      items:
        description: |-
          Items breaks the total down when more than one thing was mentioned. Empty
          when only one thing was mentioned; strict structured-output schemas
          require every property to be present, so this can't be omitempty.
        items:
          $ref: '#/definitions/domain.ExtractedExpenseItem'
        type: array
      occurred_at:
        description: OccurredAt is when the expense happened.
        type: string
      title:
        description: Title is a short human title for the expense.
        type: string
    type: object
  domain.ExtractedExpenseItem:
    properties:
      amount_minor:
        description: AmountMinor is the item's amount in minor currency units.
        type: integer
      title:
        description: Title is what this line item was for.
        type: string
    type: object
...
  handler.budgetRequest:
    properties:
      amount_limit_minor:
        type: integer
      category_id:
        type: string
      currency:
        type: string
      due_at:
        type: string
      name:
        type: string
      period:
        $ref: '#/definitions/domain.BudgetPeriod'
      shared_group_id:
        type: string
      threshold_percent:
        type: integer
    type: object
...
  handler.dictateResponse:
    properties:
      draft:
        $ref: '#/definitions/domain.ExtractedExpense'
      transcript:
        type: string
    type: object
...
  handler.intelligentRequest:
    properties:
      text:
        type: string
    type: object
...
info:
  contact: {}
  description: API for personal cost tracking.
  title: BillPiggy API
  version: "1.0"
paths:
...
  /billpiggy/api/v1/budgets/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.BudgetRecord'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: List budgets
      tags:
      - budgets
    post:
      consumes:
      - application/json
      parameters:
      - description: Budget
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.budgetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.BudgetRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create budget
      tags:
      - budgets
  /billpiggy/api/v1/budgets/{budgetID}:
    delete:
      parameters:
      - description: Budget ID
        in: path
        name: budgetID
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete budget
      tags:
      - budgets
    get:
      parameters:
      - description: Budget ID
        in: path
        name: budgetID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.BudgetRecord'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get budget
      tags:
      - budgets
    put:
      consumes:
      - application/json
      parameters:
      - description: Budget ID
        in: path
        name: budgetID
        required: true
        type: string
      - description: Budget
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.budgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.BudgetRecord'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update budget
      tags:
      - budgets
...
  /billpiggy/api/v1/expenses/dictate:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Audio recording, maximum 10 MiB
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.dictateResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Dictate a draft expense
      tags:
      - expenses
  /billpiggy/api/v1/expenses/intelligent:
    post:
      consumes:
      - application/json
      parameters:
      - description: Description
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.intelligentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ExtractedExpense'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Parse a sentence into a draft expense
      tags:
      - expenses
  /billpiggy/api/v1/expenses/scan:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Receipt image, maximum 10 MiB
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ExtractedExpense'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Scan a receipt into a draft expense
      tags:
      - expenses
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and the access token from POST
      /auth/login or POST /auth/refresh.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"